+++ pumpkin-cli/src/flatzinc/compiler/mod.rs
@@ -1,5 +1,6 @@
 mod collect_domains;
 mod context;
+mod create_objective;
 mod define_constants;
 mod define_variable_arrays;
 mod merge_equivalences;
@@ -25,8 +26,10 @@ pub fn compile(
     collect_domains::run(&ast, &mut context)?;
     define_variable_arrays::run(&ast, &mut context)?;
     post_constraints::run(&ast, &mut context)?;
+    let objective_function = create_objective::run(&ast, &mut context)?;
 
     Ok(FlatZincInstance {
         outputs: context.outputs,
+        objective_function,
     })
 }